Start of the major refactor - split all classes into separate files
